- made several unnecessary public methods protected - made UnitDisk's calcInterference method useable in initialisation stage 0

+++ modules/connectionManager/UnitDisk.cc
@@ -9,7 +9,7 @@ void UnitDisk::initialize(int stage)
 	{
 		//cPar *p = addPar("logName"); //TODO: addPar removed in omnet 4, replace it
 		//p->setStringValue("UnitDisk");
-		radioRange = par("radioRange");
+		radioRange = par("radioRange").doubleValue();
 		EV << "UnitDisk initialised with range "<<radioRange<<endl;
 	}
 }
